chore: add prettier-plugin-solidity (#833)

+++ packages/smart-contracts/src/contracts/test/BytesUtilsMock.sol
@@ -1,12 +1,12 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.0;
 
-import "../Bytes.sol";
-
+import '../Bytes.sol';
 
 contract BytesUtilsMock {
   bytes32 public extractBytes32Result;
-  function extractBytes32(bytes memory data, uint offset) public {
+
+  function extractBytes32(bytes memory data, uint256 offset) public {
     extractBytes32Result = Bytes.extractBytes32(data, offset);
   }
 }
